refactor: split algo.c into per-algorithm files under src/algos/
Break the ~3800-line src/algo.c (40 transform functions across 11
algorithm groups) into per-algorithm files for better navigability.

New structure under src/algos/:
- algos.h — forward declarations of all 40 algo functions
- lossless.c — algo_{encode,decode}_lossless
- cast.c — cast32_64d, cast16_32f, cast16_64d encode+decode
- log2.c — log_2_transform 32f/64d encode+decode
- delta.c — delta16/24/32 transform 32f/64d encode+decode
- vdelta.c — vdelta16/24 transform 32f/64d encode+decode
- vbr.c — vbr 32f/64d encode+decode
- bitpack.c — bitpack 32f/64d encode+decode

src/algo.c is trimmed to only the 3 dispatcher functions
(set_compress_algo, set_decompress_algo, get_algo_type) with
an #include of algos/algos.h.

Build scripts updated: cli/CMakeLists.txt, node-ts/CMakeLists.txt,
node/CMakeLists.txt, python/setup.py — all now glob src/algos/*.c.

Closes #121
